Fixed bug in MockCanvasManager that was not passing the frame count to shots
Fixed lack of FPS estimation for testcases.  It is now properly estimated.

Refactored a couple places having to do with shot detection

Fixed a couple bugs with how motion / brightness were overlayed on frames

Tweaked shot detection and added an additional metric (density) to filter more false positives

Reduced threshold for deduplication.  It is now 5% instead of 10%.

I made brightness not a failure case - It is still a warn-the-user case but it is not a complete shot detection disabler