TossUP/CanToss
Targets are "thrown" into the air and come back down.  This was an
effort to create a speed drill and partially satisfy the "duck hunt"
issue.

The motion of the targets that are thrown in the air causes false green
shots.  I currently ignore green lasers in the preferences.

There is significant lag when targets are moving fast.  I replicated
this in bouncing targets as well.  A shot on the trailing edge of the
fast moving target will not produce a hit.  To compensate for this, I
have an ellipse coupled with the image in the target file so when the
user hits the gif on the screen, the lag will show the hit in the
ellipse and the user gets credit for it.

To create the "report card of hits" at the end of a round, I needed a
way to get the location of the shot when it was found in the hit region
of the target.  This is already done in the canvasManager so I added a
regionImpact x and y in the interface of the targetRegion along with
gets and sets so the canvasManager can record the impact coordinates in
the TargetRegion.  If the shotListener was provided the Hit instead of
only the targetRegion, this would not be necessary.  My changes only
require changes to the regions, not the existing training exercises
(like a new overloaded shotListener would require) so I did not overload
the shotListener to return the Hit instead.  Either way, I would get
what I needed.