Space Invaders
This projector exercise is an adaptation of Space Invaders, the video
game from the 80's.  So it turns out that the moving blaster canon has
malfunctioned and you have to shoot the aliens with your laser trainer.
Who would have thought the aliens are vulnerable to 5mw of red light...
The top row of aliens are worth 15 points.  The middle two rows are
worth 10 points.  The bottom two rows are worth 5 points. The UFO is
worth 20 points.  Firing and not hitting anything will take away 1 point
per shot.  Win the game by hitting all of the aliens as fast as you can
but beware...the aliens will speed up when they feel like they are
threatened.  If the aliens hit the barriers at the bottom of the screen,
the game is over and the world as you know it will cease to
exist...until you press the reset button.